Correct BUILD_NUM value that was passed $DEBVER Add date file as reference for Changelog.Debian files Do not refetch APCu source if it exists localy Reduce apcu.sh logs Move mkpkg logs to .latest_logs subdir of shared 'dist' directory Add make test and log, with copy of report in logs subdir Log rmed files Add pear.conf to cli conffiles Correct Changelog.Debian format Add --disable-cli, --disable-cgi, --disable-phpdbg to mod/mkpkg Update tmp/.gitignore Update NOTES.md and PLANNING.md to reflect our progress Add TODO_NEXT.md

+++ php/7/Dockervars.sh
@@ -3,12 +3,13 @@
 #
 Dir=php/7
 
-echo "Fetching APCu extension..."
 PECLURL="http://pecl.php.net/get/APCu"
-rm -f $Dir/apcu-*.tgz
 Tgz=`curl -sISL "$PECLURL" | sed -n 's/^Content-Disposition:.*filename=\(.*\)$/\1/p'`
-curl -sSL "$PECLURL" -o $Dir/$Tgz
-
+if [ ! -f $Dir/$Tgz ]; then
+    rm -f $Dir/apcu-*.tgz
+    echo "Fetching APCu `expr $Tgz : 'apcu-\(.*\).tgz'` extension..."
+    curl -sSL "$PECLURL" -o $Dir/$Tgz
+fi
 EXTCOPY="COPY $Dir/$Tgz $BUILD_TOP/files
 RUN mkdir $BUILD_TOP/hooks
 COPY $Dir/apcu.sh $BUILD_TOP/hooks"
